Flexible OkHttpClient/Retrofit injection
Currently we only have 1 web API interface so this change may be cumbersome,
but now we have a flexibility to customize and reuse OkHttpClient/Retrofit instances.

For OkHttpClient, each OkHttpClient instance has its own connection pool as of OkHttp3 so OkHttpClient instance should be singleton.
If you need more configuration based on the singleton instance, you can do so using OkHttpClient#newBuilder().

Same applies to Retrofit.

Stetho has an interceptor for OkHttp3 to analyze HTTP requests/responses so this commit installs it only on debug apk.